requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.32
alembic==1.13.2
psycopg2-binary==2.9.9
pydantic-settings==2.4.0
PyJWT==2.9.0
bcrypt==4.2.0
httpx==0.27.2
python-dotenv==1.0.1
jinja2==3.1.4
python-multipart==0.0.9
pytest==8.3.2
